1002 bcf unmirror

+++ poseidon/controllers/bcf/bcf.py
@@ -298,18 +298,18 @@ class BcfProxy(JsonMixin, CookieAuthControllerProxy):
         return retval
 
     def get_seq_by_mac(self, mac):
-        my_filter = self.get_span_fabric().get('filter')
         retval = []
-        if my_filter is not None:
-            for f in my_filter:
-                # TODO check this to make sure it still works as expected
-                if 'match-specification' in f:
-                    dst = f[
-                        'match-specification'].get('dst-mac', 'broke')
-                    src = f[
-                        'match-specification'].get('src-mac', 'broke')
-                    if src == mac or dst == mac:
-                        retval.append(f.get('seq'))
+        my_filter = self.get_span_fabric().get('filter')
+        endpoints = self.get_bymac(mac)
+        self.logger.debug('Endpoints found: {0}'.format(endpoints))
+        for endpoint in endpoints:
+            if 'attachment-point' in endpoint and 'switch-interface' in endpoint['attachment-point']:
+                interface = endpoint['attachment-point']['switch-interface'].get('interface')
+                switch = endpoint['attachment-point']['switch-interface'].get('switch')
+                if my_filter is not None:
+                    for f in my_filter:
+                        if f.get('interface') == interface and f.get('switch') == switch:
+                            retval.append(f.get('seq'))
         return retval
 
     def mirror_mac(self, mac, switch, port, messages=None):
@@ -337,6 +337,7 @@ class BcfProxy(JsonMixin, CookieAuthControllerProxy):
     def unmirror_mac(self, mac, switch, port, messages=None):
         status = None
         kill_list = self.get_seq_by_mac(mac)
+        if not kill_list: return True
         for kill in kill_list:
             self.logger.debug('unmirroring: {0}'.format(kill))
             self.mirror_traffic(kill, mirror=False)
@@ -389,6 +390,9 @@ class BcfProxy(JsonMixin, CookieAuthControllerProxy):
           }
 
         '''
+        if not mirror:
+            self.logger.debug("Attempting to unmirror")
+        
         resource = fabric_span_endpoint.format(self.span_fabric_name)
         uri = urljoin(self.base_uri, resource)
         data = self.get_span_fabric()  # first element is vent span rule
@@ -407,6 +411,7 @@ class BcfProxy(JsonMixin, CookieAuthControllerProxy):
         else:  # mirror=False
             data['filter'] = [filter for filter in data[
                 'filter'] if filter['seq'] != seq]
+            self.logger.debug("unmirror put body: {0}".format(data))
         r = self.request_resource(method='PUT', url=uri, data=json.dumps(
             data), verify=(not self.trust_self_signed_cert))
         retval = BcfProxy.parse_json(r)
